tools/levels_engine/requirements.txt
# requirements for levels_engine (add to global venv)
python-dotenv>=1.0
requests>=2.28
pandas>=2.0
pydantic>=2.0
numpy>=1.24
beautifulsoup4>=4.12
pytest>=7.0
